Kernel: add limits.h with INT_MIN and INT_MAX definitions; implement chdir and getcwd syscalls; refactor mkdir syscall return type; enhance shell commands with cd and pwd; improve keyboard input handling; add path collapsing utility

+++ module/src/libc/sys/stat/mkdir.c
@@ -4,7 +4,7 @@
 #include "sys/syscall.h"
 
 int mkdir(const char *pathname, mode_t mode) {
-    ssize_t res = syscall_mkdir(pathname, mode);
+    int res = syscall_mkdir(pathname, mode);
     if (res < 0) {
         errno = -res;
         return -1;
